Add support for item extra hints
This commit adds support for a new `extra` field in the `item` struct,
which allows extra information to be displayed for each item in the
menu.

The extra hint is separated from the menu item using a configurable
separator (`hint_separator`), which can be set using the new `-h`
option.

Drawing of the extra hint was added to `drawitem()`. The rightmost side
of the item text area is now used for the hint field.